+++ Clutch/FrameworkDumper.m
@@ -63,12 +63,12 @@
 
     for (int i = 0; i < _thinHeader.header.ncmds; i++) {
 
-        uint32_t cmd = [newFileHandle intAtOffset:newFileHandle.offsetInFile];
-        uint32_t size = [newFileHandle intAtOffset:newFileHandle.offsetInFile+sizeof(uint32_t)];
+        uint32_t cmd = [newFileHandle intAtOffset:(NSUInteger)newFileHandle.offsetInFile];
+        uint32_t size = [newFileHandle intAtOffset:(NSUInteger)newFileHandle.offsetInFile+sizeof(uint32_t)];
 
         switch (cmd) {
             case LC_CODE_SIGNATURE: {
-                [newFileHandle getBytes:&ldid inRange:NSMakeRange(newFileHandle.offsetInFile,sizeof(struct linkedit_data_command))];
+                [newFileHandle getBytes:&ldid inRange:NSMakeRange((NSUInteger)newFileHandle.offsetInFile,sizeof(struct linkedit_data_command))];
                 foundSignature = YES;
 
                 [[ClutchPrint sharedInstance] printDeveloper: @"FOUND CODE SIGNATURE: dataoff %u | datasize %u",ldid.dataoff,ldid.datasize];
@@ -76,8 +76,8 @@
                 break;
             }
             case LC_ENCRYPTION_INFO: {
-                cryptlc_offset = newFileHandle.offsetInFile;
-                [newFileHandle getBytes:&crypt inRange:NSMakeRange(newFileHandle.offsetInFile,sizeof(struct encryption_info_command))];
+                cryptlc_offset = (uint32_t)newFileHandle.offsetInFile;
+                [newFileHandle getBytes:&crypt inRange:NSMakeRange((NSUInteger)newFileHandle.offsetInFile,sizeof(struct encryption_info_command))];
                 foundCrypt = YES;
 
                 [[ClutchPrint sharedInstance] printDeveloper: @"FOUND ENCRYPTION INFO: cryptoff %u | cryptsize %u | cryptid %u",crypt.cryptoff,crypt.cryptsize,crypt.cryptid];
@@ -86,7 +86,7 @@
             }
             case LC_SEGMENT:
             {
-                [newFileHandle getBytes:&__text inRange:NSMakeRange(newFileHandle.offsetInFile,sizeof(struct segment_command))];
+                [newFileHandle getBytes:&__text inRange:NSMakeRange((NSUInteger)newFileHandle.offsetInFile,sizeof(struct segment_command))];
 
                 if (strncmp(__text.segname, "__TEXT", 6) == 0) {
                     foundStartText = YES;
@@ -118,7 +118,7 @@
     codesignblob = malloc(ldid.datasize);
 
     [newFileHandle seekToFileOffset:_thinHeader.offset + ldid.dataoff];
-    [newFileHandle getBytes:codesignblob inRange:NSMakeRange(newFileHandle.offsetInFile, ldid.datasize)];
+    [newFileHandle getBytes:codesignblob inRange:NSMakeRange((NSUInteger)newFileHandle.offsetInFile, ldid.datasize)];
 
     [[ClutchPrint sharedInstance] printDeveloper: @"hello it's me"];
